requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
deprecation==2.1.0
gotrue==2.4.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
outcome==1.3.0.post0
packaging==24.1
postgrest==0.16.8
pydantic==2.7.4
pydantic_core==2.18.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.6
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.1
websockets==12.0
wsproto==1.2.0
webdriver-manager==3.8.6

tweepy~=4.14.0